+++ crypto_sign/faest_em_128f/m4/crypto_sign.h
@@ -9,12 +9,13 @@
 #define CRYPTO_PUBLICKEYBYTES 32
 #define CRYPTO_BYTES 5696
 #define CRYPTO_ALGNAME "faest_em_128f"
+#include <stddef.h>
 
 int crypto_sign_keypair(unsigned char* pk, unsigned char* sk);
-int crypto_sign(unsigned char* sm, unsigned long long* smlen, const unsigned char* m,
-                unsigned long long mlen, const unsigned char* sk);
-int crypto_sign_open(unsigned char* m, unsigned long long* mlen, const unsigned char* sm,
-                     unsigned long long smlen, const unsigned char* pk);
+int crypto_sign(unsigned char* sm, size_t* smlen, const unsigned char* m,
+                size_t mlen, const unsigned char* sk);
+int crypto_sign_open(unsigned char* m, size_t* mlen, const unsigned char* sm,
+                     size_t smlen, const unsigned char* pk);
 
 #endif
